[core] Rewrite of stable ELPA repository support
* rename variable dotspacemacs-use-spacelpa to dotspacemacs-use-stable-elpa
* rename variable dotspacemacs-verify-spacelpa-archives to
  dotspacemacs-verify-stable-elpa-archives
* new lock files: .lock-stable and .lock-latest

On master branch the usage of .lock-stable is now mandatory. On develop branch
the file .lock-latest is used unless the variable dotspacemacs-use-stable-elpa
is set to t in which case the develop branch behaves like the master branch.
This is useful for tested the stable ELPA repository on the develop branch.

The location of the stable ELPA repositoy has been moved from .cache/stable-elpa
directory to the simpler path .emacs.d/elpa/repository. It is now possible to
drop the manually downloaded stable ELPA files in this directory and Spacemacs
will install it automatically.

Use curl or wget to download the files instead of built-in Emacs function which
causes troubles with large files sometimes.

Fix the import of the Spacemacs GPG public key.